refactor(srv): split blockcontroller/shell.rs into submodules
