+++ .changesets/1782973832-refactor-srv-split-shell-controller-into-submodules-8ksv.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(srv): split shell controller into submodules
